🎨 Style: 공통 컴포넌트 최상위 위치하도록 _추가
